latihan program luas lingkaran

+++ Day024.java
@@ -7,7 +7,7 @@ public class main {
         Scanner p = new Scanner (System.in);
         
         System.out.print("masukkan angka: ");
-        double  a = p.nextInt();
+        double  a = p.nextDouble();
         
         double luas = Math.PI * a * a;
         System.out.println("Luas Lingkaran = " + luas);
